@@ -6,7 +6,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.0'
+        classpath 'com.android.tools.build:gradle:3.5.2'
     }
 }
 
@@ -20,10 +20,10 @@ repositories {
 dependencies {
 
 
-        implementation "com.android.support:support-v4:28.0.0"
-        implementation "com.android.support:support-v13:28.0.0"
-        implementation "com.android.support:cardview-v7:28.0.0"
-        implementation "com.android.support:appcompat-v7:28.0.0"
+        implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+        implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+        implementation 'androidx.cardview:cardview:1.0.0'
+        implementation 'androidx.appcompat:appcompat:1.0.0'
 
 
 
@@ -41,11 +41,11 @@ List<String> dirs = [
     'template'] // boilerplate code that is generated by the sample template process
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 28
+        minSdkVersion 29
+        targetSdkVersion 29
     }
 
     compileOptions {
